Workaround found! 2 minor bugs left to fix!
After much trial and error I found _.debounce() from the underscore.js library to fix my problem. I might simplify this later into my own plug in to remove the underscore.js dependancy but it might be something we need later.

To elaborate a tad, underscore.js is like a giant freaking library of functions that can be added to a project to help minimize the amount of user-defined functions you need to write

It took a couple of hours but it's pretty straight forward to get acclimated.

Taking a short break then I'll be moving on.