Multiple changes incomming
Several file renames to remove the  naming scheme since we are out moving towards production, this meant renaming parts of all the views and the controller. I updated the project name to be  since our build has 3 different projects to make it more descriptive. I also got the PRG to work for the CMS. So  when logged in and  when logged out. I still need to style the view and better seperate the content